+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - FormMultiTextField Component
 - Consequential Call Out Examples
 - Exposed After Pattern Examples
+- Pagination Component
 
 ### Changed